Refactor: Reorganized naming to remove interview test association
- Renamed classes, packages, and identifiers to reflect project purpose
- Removed references tied to the original interview context
- Updated documentation and comments for clarity
