+++ src/main/java/io/juanez/cryptosim/entity/HoubiEntity.java
@@ -1,4 +1,4 @@
-package com.dotjuan89.aquariuxtest.entity;
+package io.juanez.cryptosim.entity;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
