+++ src/main/java/io/juanez/cryptosim/repository/PricesRepository.java
@@ -1,6 +1,6 @@
-package com.dotjuan89.aquariuxtest.repository;
+package io.juanez.cryptosim.repository;
 
-import com.dotjuan89.aquariuxtest.entity.PricesEntity;
+import io.juanez.cryptosim.entity.PricesEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
