+++ src/main/java/io/juanez/cryptosim/repository/UsersRepository.java
@@ -1,6 +1,6 @@
-package com.dotjuan89.aquariuxtest.repository;
+package io.juanez.cryptosim.repository;
 
-import com.dotjuan89.aquariuxtest.entity.UsersEntity;
+import io.juanez.cryptosim.entity.UsersEntity;
 import org.hibernate.sql.Update;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
