+++ src/main/resources/application.properties
@@ -2,7 +2,7 @@
 trade.verbose=true
 
 # H2
-spring.datasource.url=jdbc:h2:mem:aquariux
+spring.datasource.url=jdbc:h2:mem:cryptosim
 spring.datasource.driverClassName=org.h2.Driver
 spring.datasource.username=sa
 spring.datasource.password=password
